Move canvas template to Canvas
