+++ src/App.css
@@ -26,9 +26,6 @@
 
 .text-control input {
   margin-bottom: 1rem;
-}
-
-.text-control input {
   max-width: 100%;
 }
