--- src/my-canvas-element/index.ts
@@ -1,3 +1,0 @@
-import { MyCanvasElement } from './myCanvasElement';
-
-window.customElements.define('my-canvas-element', MyCanvasElement);
